Added remove handler and other changes

+++ ToDoApi/Migrations/ToDoContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace ToDoApi.Migrations
 
             modelBuilder.Entity("ToDoApi.Models.Comment", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Content")
                         .IsRequired()
@@ -38,8 +38,8 @@ namespace ToDoApi.Migrations
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
 
-                    b.Property<int>("ToDoItemId")
-                        .HasColumnType("int");
+                    b.Property<long>("ToDoItemId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
 
@@ -50,11 +50,11 @@ namespace ToDoApi.Migrations
 
             modelBuilder.Entity("ToDoApi.Models.Tag", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -68,11 +68,11 @@ namespace ToDoApi.Migrations
 
             modelBuilder.Entity("ToDoApi.Models.ToDoDetails", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<int>("EstimatedMinutes")
                         .HasColumnType("int");
@@ -85,8 +85,8 @@ namespace ToDoApi.Migrations
                         .HasMaxLength(1000)
                         .HasColumnType("nvarchar(1000)");
 
-                    b.Property<int>("ToDoItemId")
-                        .HasColumnType("int");
+                    b.Property<long>("ToDoItemId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
 
@@ -98,11 +98,11 @@ namespace ToDoApi.Migrations
 
             modelBuilder.Entity("ToDoApi.Models.ToDoItem", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
@@ -136,11 +136,11 @@ namespace ToDoApi.Migrations
 
             modelBuilder.Entity("ToDoItemTag", b =>
                 {
-                    b.Property<int>("TagId")
-                        .HasColumnType("int");
+                    b.Property<long>("TagId")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("ToDoItemId")
-                        .HasColumnType("int");
+                    b.Property<long>("ToDoItemId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("TagId", "ToDoItemId");
