+++ ToDoApi/Models/Comment.cs
@@ -2,10 +2,10 @@
 
 public class Comment
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Content { get; set; } = null!;
     public DateTime CreatedAt { get; set; }
 
-    public int ToDoItemId { get; set; }
+    public long ToDoItemId { get; set; }
     public ToDoItem ToDoItem { get; set; } = null!;
 }
