+++ ToDoApi/Models/ToDoDetails.cs
@@ -2,12 +2,12 @@
 
 public class ToDoDetails
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     public string? Notes { get; set; }
     public string? Location { get; set; }
     public int EstimatedMinutes { get; set; }
 
-    public int ToDoItemId { get; set; }
+    public long ToDoItemId { get; set; }
     public ToDoItem ToDoItem { get; set; } = null!;
 }
